Fix #66: laserCavity test wip + rename validator
