+++ tests/propagation_test.py
@@ -62,17 +62,10 @@ def test_propagation():
     lc.propagate(num_cycles=4, callback=intensity_callback)
     pkeq(
         str([
-            [-0.09007729696643918, -0.07369960660890479, -0.05732191625137039, -0.04094422589383599,
-             -0.0245665355363016, -0.00818884517876721, 0.008188845178767196, 0.024566535536301587,
-              0.04094422589383598, 0.05732191625137037, 0.07369960660890476], [0.0004206083571113905,
-              0.00042148866696306385, 0.0004223794433054218, 0.0004232805397432438, 0.0004241917784428871,
-              0.000425113299639614, 0.00042604485049852994, 0.0004269865220384565, 0.0004279381999672457,
-              0.00042889980587495087, 0.0004298713099470023], [0.0004206082780286289, 0.00042148868408716775,
-              0.00042237944785462596, 0.00042328054426942787, 0.00042419175847993286, 0.0004251132953753882,
-               0.00042604485714848017, 0.0004269865118091315, 0.0004279381772178525, 0.00042889976698366305,
-               0.00042987128270201806], [768161600000.0, 3386752600000.0, 11088761000000.0, 26961980000000.0,
-               48684415000000.0, 65282614000000.0, 65009240000000.0, 48075314000000.0, 26402127000000.0,
-               10767783000000.0, 3261246500000.0], [1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55]
-               ]),
+            [-0.09007729696643918, -0.07369960660890479, -0.05732191625137039, -0.04094422589383599, -0.0245665355363016, -0.00818884517876721, 0.008188845178767196, 0.024566535536301587, 0.04094422589383598, 0.05732191625137037, 0.07369960660890476],
+            [0.00042060832381247547, 0.00042148869279589853, 0.00042237938068257706, 0.00042328053915324966, 0.0004241917709675302, 0.000425113259460114, 0.0004260448258076334, 0.0004269864714645692, 0.00042793823411483965, 0.0004288998097451079, 0.00042987124363644844],
+            [0.000420608251804938, 0.00042148867911241705, 0.00042237936420157245, 0.00042328052228606146, 0.0004241917082775849, 0.0004251133507806255, 0.00042604486544507405, 0.00042698643452929934, 0.00042793819893895553, 0.00042889977737009993, 0.00042987124878598516],
+            [768158600000.0, 3386736400000.0, 11088713000000.0, 26961888000000.0, 48684220000000.0, 65282380000000.0, 65008927000000.0, 48075112000000.0, 26401999000000.0, 10767730000000.0, 3261234200000.0],
+            [1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55, 1.55]]),
         str(results[-1]),
     )
